requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
fonttools==4.33.3
idna==3.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.4.3
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.22.4
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pylab-sdk==1.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
stack-data==0.3.0
threadpoolctl==3.1.0
toml==0.10.2
traitlets==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
